+++ examples/phpcs-standard-package/Compatibility/Sniffs/Arrays/ArraySniff.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Acme\Compatibility\Sniffs\Arrays;
+
+class ArraySniff
+{
+}